backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pymongo==4.6.0
motor==3.3.2
langchain==0.1.0
langchain-google-genai==1.0.0
langchain-mongodb==0.1.0
pydantic-ai==0.0.12
openai==1.3.0
google-generativeai==0.3.0
Pillow==10.1.0
PyPDF2==3.0.1
sentence-transformers==2.2.2
numpy==1.24.3
httpx==0.25.2
pydantic==2.5.0
email-validator==2.1.0